Updating the formatting of the request dictionary
We found that when given 'data' the user has to put the dictionary inside single quotes.  This is required for the Requests library to get the proper data back.  EG: {'data': '{"username": "username"}'}
We also had to remove the spaces in 'Content-Type' because it has different behavior based on whether or not there are spaces there.  Added in a test for searching for 'apps'.  This is the most we can do for now because we can't do system tests because that would require us to add a username and password to Git.